refactor: move JumpTo and related api methods to eregs-vite
