+++ solution/ui/regulations/eregs-component-lib/src/components/index.js
@@ -10,7 +10,6 @@ export { default as DropdownHeader } from "./DropdownHeader.vue";
 export { default as DropdownItem } from "./DropdownItem.vue";
 export { default as FlashBanner } from "./FlashBanner.vue";
 export { default as IFrameContainer } from "./IFrameContainer.vue";
-export { default as JumpTo } from "./JumpTo.vue";
 export { default as LastParserSuccessDate } from "./LastParserSuccessDate.vue";
 export { default as PartButton } from "./PartButton.vue";
 export { default as PrintBtn } from "./PrintBtn.vue";
