+++ solution/ui/regulations/eregs-component-lib/src/main.js
@@ -15,7 +15,6 @@ import {
     GovInfoLinks,
     HeaderComponent,
     IFrameContainer,
-    JumpTo,
     LastParserSuccessDate,
     PartButton,
     PrintBtn,
@@ -193,7 +192,6 @@ function main() {
             GovInfoLinks,
             HeaderComponent,
             IFrameContainer,
-            JumpTo,
             LastParserSuccessDate,
             PartButton,
             PrintBtn,
