+++ solution/ui/regulations/eregs-vite/src/components/JumpTo.vue
@@ -64,7 +64,7 @@
     </div>
 </template>
 <script>
-import { getTitles, getParts } from "../api";
+import { getTitles, getParts } from "../utilities/api.js";
 
 export default {
     name: "JumpTo",
@@ -145,7 +145,7 @@ export default {
 
     methods: {
         async getParts(title) {
-            const partsList = await getParts(this.apiurl, title);
+            const partsList = await getParts(title, this.apiUrl);
             this.filteredParts = partsList.map((part) => part.name);
         },
         getLink(e) {
