+++ solution/ui/regulations/eregs-vite/vite.config.js
@@ -25,7 +25,7 @@ export default defineConfig({
             legacy: path.resolve(__dirname, "../../regulations"),
             eregsComponentLib: path.resolve(
                 __dirname,
-                "../../regulations/eregs-component-lib"
+                "../eregs-component-lib"
             ),
             "@": path.resolve(__dirname, "src"),
         },
